kadai11&12

+++ src/main/java/jp/te4a/spring/boot/myapp11/BookController.java
@@ -1,9 +1,11 @@
-package jp.te4a.spring.boot.myapp9;
+package jp.te4a.spring.boot.myapp11;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -28,10 +30,14 @@ public class BookController {
 	    return "books/list";
 	  }
 	  @PostMapping(path="create")
-	  String create(BookForm form, Model mode) {
+	  String create(@Validated BookForm form, BindingResult result , Model model) {
+	    if(result.hasErrors()) {
+	      return list(model);
+	    }
 	    bookService.create(form);
 	    return "redirect:/books";
-	  }	
+	  }
+
 	  @PostMapping(path = "edit", params = "form")
 	  String editForm(@RequestParam Integer id, BookForm form) {
 	    BookForm bookForm = bookService.findOne(id);
@@ -40,10 +46,15 @@ public class BookController {
 	  }
 
 	  @PostMapping(path = "edit")
-	  String edit(@RequestParam Integer id, BookForm form) {
-	    bookService.update(form);
-	    return "redirect:/books";
+	  String edit(@RequestParam Integer id, @Validated BookForm form,
+	                                                                                                                           BindingResult result) {
+	  if(result.hasErrors()) {
+		return editForm(id, form);
 	  }
+	  bookService.update(form);
+	  	return "redirect:/books";
+	  }
+
 	  @PostMapping(path = "delete")
 	  String delete(@RequestParam Integer id) {
 	    bookService.delete(id);
